Description :
- User is allowed to enter concentration and quantity fields on specimen pages - add, edit, derivative, aliquot, BO and
  Specimen requirement
- In specimen concentration and quantity field, any number entered more than 1 million is converted into scientific notation
- In specimen edit, overview, the field value is be displayed in scientific notations.

+++ WEB-INF/src/com/krishagni/catissueplus/core/biospecimen/domain/DerivedSpecimenRequirement.java
@@ -11,7 +11,7 @@ public class DerivedSpecimenRequirement {
 	
 	private BigDecimal quantity;
 	
-	private Double concentration;
+	private BigDecimal concentration;
 	
 	private String labelFmt;
 	
@@ -53,11 +53,11 @@ public class DerivedSpecimenRequirement {
 		this.quantity = quantity;
 	}
 
-	public Double getConcentration() {
+	public BigDecimal getConcentration() {
 		return concentration;
 	}
 
-	public void setConcentration(Double concentration) {
+	public void setConcentration(BigDecimal concentration) {
 		this.concentration = concentration;
 	}
